fixed bugs when starting for the first time

+++ libs/sequelize/migrations/01-event-is-paid.coffee
@@ -1,12 +1,18 @@
 module.exports = { 
 
-  up: (knex)->    
-    knex.schema.table 'Event', (table)->
-      table.dateTime("paid_at")
+  up: (knex)->   
+    knex.schema.hasTable('Event').then (exists)->
+      if not exists then return     
+
+      knex.schema.table 'Event', (table)->
+        table.dateTime("paid_at")
   
   
   down: (knex)->
-    knex.schema.table 'Event', (table)->
-      table.dropColumn("paid_at")
+    knex.schema.hasTable('Event').then (exists)->
+      if not exists then return
+
+      knex.schema.table 'Event', (table)->
+        table.dropColumn("paid_at")
 
 }
